data_prep/pyproject.toml
[tool.poetry]
name = "data-prep"
version = "0.1.0"
description = "Data Preparation Kubeflow Component"
authors = ["Martynas Subonis <martynas.subonis@gmail.com>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "~3.12"
kfp = "2.8.0"
google-cloud-storage = "2.18.2"
scikit-learn = "1.5.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
